llm-resume-analyzer-main/llm-resume-analyzer-main/requirements.txt
streamlit
PyPDF2
spacy
openai
python-dotenv
